operators/detect_text_in_image_tesseract/pyproject.toml
[project]
name = "feluda-detect-text-in-image-tesseract"
version = "0.1.3"
requires-python = ">=3.10"
dependencies = [ "pytesseract>=0.3.10",
    "Pillow>=11.1.0",
    "requests>=2.28.0",]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.semantic_release]
version_variable = ["pyproject.toml:project.version"]

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
tag_format = "{name}-{version}"

[tool.hatch.build.targets.wheel]
packages = ["."]
